# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  docs.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      success:
        type: boolean
      timestamp:
        type: integer
    type: object
...
  docs.StandardResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      success:
        type: boolean
      timestamp:
        type: integer
    type: object
...
host: router.yeying.pub
info:
  contact: {}
  description: Router public/admin API (JWT/UCAN) under /api/v1.
  title: Router API
  version: v1
paths:
...
  /api/v1/admin/channel/test:
    get:
      parameters:
      - description: Scope (all|enabled|disabled)
        in: query
        name: scope
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/docs.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Test all channels (admin)
      tags:
      - admin
  /api/v1/admin/channel/test/{id}:
    get:
      parameters:
      - description: Channel ID
        in: path
        name: id
        required: true
        type: string
      - description: Model name
        in: query
        name: model
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/docs.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Test channel (admin)
      tags:
      - admin
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"